Changes child mort, root finding, disease_states
Adds parameters that the user can toggle to child mortlity, primarily
for the purpose of testing the access of the mother parameters from the
child.
Replaces the root finding procedures with a simple inverted version of
the function.
Removes the infected agent_var and replaces it with the more general
disease_status variable